requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
certifi==2018.10.15
chardet==3.0.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pycodestyle==2.4.0
pylint==2.1.1
pytest==4.0.1
pyzmq==17.1.2
requests==2.20.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
